tree: moved PEP command to its own plugin file
