+++ mcodingbot/plugins/info.py
@@ -28,12 +28,3 @@ class LinksCommand:
             ),
         )
         await ctx.respond(embed=embed)
-
-
-@plugin.include
-@crescent.command(name="pep", description="Find a Python Enhancement Proposal")
-class PEPCommand:
-    pep_number = crescent.option(int, "the PEP number")
-
-    async def callback(self, ctx: crescent.Context) -> None:
-        await ctx.respond(f"https://peps.python.org/pep-{self.pep_number:03}/")
